update: change the sorting logic based on 'vote_average'
The sorting logic before we used based on 'vote_average' included movies with just one rating count, so movies that no one has ever heard of are displayed at the top of the list with a score of 10. So changed the rating(vote_average) to sorted in descending order among movies with more than 300 evaluations(vote_count).